move file operation tools to the shared tools plugin
