+++ plugins/enthusiast-agent-tools/src/enthusiast_agent_tools/tools/__init__.py
@@ -1,9 +1,13 @@
+from .file_list_tool import FileListTool
+from .file_retrieval_tool import FileRetrievalTool
 from .product_examples_tool import ProductExamplesTool
 from .product_sql_search_tool import ProductSQLSearchTool
 from .stop_execution_tool import StopExecutionTool
 from .upsert_product_details_tool import UpsertMemoryEntry, UpsertProductDetailsTool
 
 __all__ = [
+    "FileListTool",
+    "FileRetrievalTool",
     "ProductExamplesTool",
     "ProductSQLSearchTool",
     "StopExecutionTool",
