+++ plugins/enthusiast-common/enthusiast_common/agents/base.py
@@ -7,8 +7,6 @@ from langchain_core.tools import BaseTool
 from pydantic import BaseModel
 
 from ..injectors import BaseInjector
-from ..tools.files.list_files_tool import FileListTool
-from ..tools.files.perform_file_operation_tool import FileRetrievalTool
 from ..utils import RequiredFieldsModel, validate_required_vars
 
 
@@ -44,7 +42,6 @@ class BaseAgent(ABC, ExtraArgsClassBase):
     TOOLS = []
 
     FILE_UPLOAD = False
-    DEFAULT_FILE_TOOLS = [FileListTool, FileRetrievalTool]
 
     def __init__(
         self,
@@ -61,15 +58,6 @@ class BaseAgent(ABC, ExtraArgsClassBase):
         self._conversation_id = conversation_id
         self._callback_handler = callback_handler
         self._injector = injector
-
-    def __init_subclass__(cls, **kwargs):
-        super().__init_subclass__(**kwargs)
-        if getattr(cls, "FILE_UPLOAD", False):
-            from ..config import FileToolConfig
-
-            cls.TOOLS = getattr(cls, "TOOLS", []) + [
-                FileToolConfig(tool_class=file_tool_class) for file_tool_class in cls.DEFAULT_FILE_TOOLS
-            ]
 
     @abstractmethod
     def get_answer(self, input_text: str) -> str:
